refactor: Dice 컴포넌트 내부 moveToken 함수 커스텀훅으로 분리
- 순간이동 기능 구현시 재사용을 위해서 커스텀훅으로 분리
